refactor screen package name
