+++ core/src/blog/boomerangbeast/screens/GameScreen.java
@@ -1,4 +1,4 @@
-package blog.boomerangbeast.views;
+package blog.boomerangbeast.screens;
 
 public interface GameScreen {
     void pauseGame(Boolean pause);
