+++ core/src/blog/boomerangbeast/screens/menus/HighScoreScreen.java
@@ -1,14 +1,12 @@
-package blog.boomerangbeast.views.menus;
+package blog.boomerangbeast.screens.menus;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.ButtonGroup;
-import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.ScrollPane;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
@@ -16,57 +14,44 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.TiledDrawable;
-import com.badlogic.gdx.utils.IntMap;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
 import blog.boomerangbeast.BoomerangBeast;
+import blog.boomerangbeast.highscore.Save;
 
-public class LevelSelectionScreen implements Screen {
-    private static final int IMG_WIDTH = 700;
-    private static final int IMG_HEIGHT = 600;
-
-
+public class HighScoreScreen implements Screen {
+    private long[] highScores;
+    private String[] names;
     private BoomerangBeast parent;
-    private Stage stage;
     private Skin skin;
+    private TextureAtlas atlas;
+    private Stage stage;
     private TextureAtlas.AtlasRegion background;
-    private TextureAtlas gameAtlas;
-    private int levelSelected = 1;
-    private Image levelImage;
-    private IntMap<Image> levelPreviewImages;
-    private Table innerTable;
-    private Table outerTable;
-    private Table previewTable;
-    private int characterSelected;
+    private int levelSelected;
+    private Table outerTable, innerTable, highscoreTable;
 
 
-    public LevelSelectionScreen(BoomerangBeast boomerangBeast, int character) {
-        parent = boomerangBeast;
+    public HighScoreScreen(BoomerangBeast parent, int level){
+        this.parent = parent;
         stage = new Stage(new ScreenViewport());
-        characterSelected = character;
+        levelSelected = level;
+        highScores = Save.hsd.get(levelSelected).getHighScores();
 
-        // Get images to display loading progress
-        gameAtlas = parent.assMan.manager.get("images/game.atlas");
-
+        atlas = parent.assMan.manager.get("images/game.atlas");
         skin = parent.assMan.manager.get("skin/game/game.json");
-        background = gameAtlas.findRegion("background");
+        background = atlas.findRegion("background");
 
-        //Load level preview images
-        levelPreviewImages = new IntMap<>();
-        levelPreviewImages.put(1, new Image(new Texture("preview/level1.png")));
-        levelPreviewImages.put(2, new Image(new Texture("preview/level2.png")));
-        levelPreviewImages.put(3, new Image(new Texture("preview/level3.png")));
-        levelPreviewImages.put(4, new Image(new Texture("preview/level4.png")));
+        Save.load();
+
     }
-
-
-
 
     @Override
     public void show() {
         Gdx.input.setInputProcessor(stage);
 
-        levelSelected = 1;
+        highScores = Save.hsd.get(levelSelected).getHighScores();
+        names = Save.hsd.get(levelSelected).getNames();
+
 
         // Create text buttons, labels etc.
         final TextButton l1 = new TextButton("Homecoming", skin, "toggle2");
@@ -74,40 +59,60 @@ public class LevelSelectionScreen implements Screen {
         final TextButton l3 = new TextButton("Mountain Jam", skin, "toggle2");
         final TextButton l4 = new TextButton("Platform Plooza", skin, "toggle2");
 
-        l1.setChecked(true);
+        switch (levelSelected){
+            case 1:
+                l1.setChecked(true);
+                break;
+            case 2:
+                l2.setChecked(true);
+                break;
+            case 3:
+                l3.setChecked(true);
+                break;
+            case 4:
+                l4.setChecked(true);
+                break;
+        }
 
         ButtonGroup buttonGroup = new ButtonGroup(l1, l2, l3, l4);
         buttonGroup.setMaxCheckCount(1);
 
-        Label headerLabel = new Label("SELECT LEVEL", skin, "big");
+        Label headerLabel = new Label("HIGH SCORE", skin, "big");
 
         final TextButton backButton = new TextButton("Back", skin, "blue-small");
-        final TextButton playButton = new TextButton("Play", skin, "blue-small");
-
-
-        levelImage = levelPreviewImages.get(levelSelected);
-        levelImage.setSize(IMG_WIDTH,IMG_HEIGHT);
 
 
         innerTable = new Table();
         outerTable = new Table();
-        previewTable = new Table();
+        highscoreTable = new Table();
 
-        if (BoomerangBeast.DEBUG) {
+
+        if (BoomerangBeast.DEBUG){
             innerTable.setDebug(true);
             outerTable.setDebug(true);
-            previewTable.setDebug(true);
+            highscoreTable.setDebug(true);
         }
 
         outerTable.setFillParent(true);
-        previewTable.setFillParent(true);
-        previewTable.setBackground(new TiledDrawable(background));
+        highscoreTable.setFillParent(true);
 
+        highscoreTable.setBackground(new TiledDrawable(background));
 
-        previewTable.center().padLeft(Gdx.graphics.getWidth()/4);
-        previewTable.add(levelImage).size(levelImage.getWidth(), levelImage.getHeight()).expandX();
+        highscoreTable.clear();
+        highscoreTable.center().padLeft(Gdx.graphics.getWidth()/8);
 
+        //Add highscores to table
+        for (int i = 0; i < highScores.length; i++) {
+            Label num = new Label(String.format("%2d.", i + 1), skin, "highscore");
+            Label score = new Label(String.format("%7s", highScores[i]), skin, "highscore");
+            Label name = new Label(String.format("%s", names[i].split("-")[0]), skin, "highscore");
 
+            highscoreTable.add(num).padRight(200).right();
+            highscoreTable.add(name).padRight(90).left();
+            highscoreTable.add(score).left();
+
+            highscoreTable.row();
+        }
 
         innerTable.add(l1).padTop(30).fillX().expandX();
         innerTable.row();
@@ -115,23 +120,19 @@ public class LevelSelectionScreen implements Screen {
         innerTable.row();
         innerTable.add(l3).padTop(30).fillX().expandX();
         innerTable.row();
-        innerTable.add(l4).padTop(30).padBottom(30).fillX().expandX();;
+        innerTable.add(l4).padTop(30).padBottom(30).fillX().expandX();
+
 
         ScrollPane scrollPane = new ScrollPane(innerTable, skin);
 
-        outerTable.add(headerLabel).colspan(3).padTop(30);
+        outerTable.add(headerLabel).center().colspan(3).padTop(30);
         outerTable.row().expandX();
-        outerTable.add(scrollPane).fillY().expandY().pad(Gdx.graphics.getBackBufferHeight()/6,0,Gdx.graphics.getBackBufferHeight()/6,0);
+        outerTable.add(scrollPane).fillY().expandY().left().pad(Gdx.graphics.getBackBufferHeight()/6, Gdx.graphics.getWidth()/8, Gdx.graphics.getBackBufferHeight()/6, 0);
 
         outerTable.row().expandX();
-        outerTable.add(backButton).pad(20,0,30,0);
-        outerTable.add();
-        outerTable.add(playButton).pad(20,0,30,0);
-
-
-        stage.addActor(previewTable);
+        outerTable.add(backButton).pad(0,0,50,0);
+        stage.addActor(highscoreTable);
         stage.addActor(outerTable);
-
 
 
         // Create button listeners
@@ -160,40 +161,31 @@ public class LevelSelectionScreen implements Screen {
                 levelSelected = 4;            }
         });
 
-
-
         backButton.addListener(new ChangeListener() {
             @Override
             public void changed(ChangeEvent changeEvent, Actor actor) {
-                parent.changeScreen(BoomerangBeast.CHARACTER_SELECTION);
+                parent.changeScreen(BoomerangBeast.MENU);
 
             }
         });
 
-        playButton.addListener(new ChangeListener() {
-            @Override
-            public void changed(ChangeEvent changeEvent, Actor actor) {
-                if (parent.isOnline()) {
-                    parent.changeScreen(BoomerangBeast.MULTIPLAYER, true, levelSelected, characterSelected);
 
-                } else {
-                    parent.changeScreen(BoomerangBeast.APPLICATION, true, levelSelected, characterSelected);
-                }
-                System.out.println(levelSelected);
-            }
-        });
+
 
 
     }
 
+
+
     @Override
-    public void render(float dt) {
+    public void render(float delta) {
+        // clear the screen ready for next set of images to be drawn
         Gdx.gl.glClearColor(0f, 0f, 0f, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
 
         update();
-        // Tell our stage to do actions and draw itself
-        stage.act(Math.min(Gdx.graphics.getDeltaTime(), 1 / 30f));
+
+        stage.act();
         stage.draw();
 
     }
@@ -202,14 +194,32 @@ public class LevelSelectionScreen implements Screen {
     public void resize(int width, int height) {
         // Change the stage's viewport when the screen size is changed
         stage.getViewport().update(width, height, true);
-
     }
 
-    private void update(){
-        previewTable.clear();
-        levelImage = levelPreviewImages.get(levelSelected);
-        levelImage.setSize(IMG_WIDTH,IMG_HEIGHT);
-        previewTable.add(levelImage).size(levelImage.getWidth(), levelImage.getHeight()).expandX();
+    public void update(){
+
+        //highscorePane.clear();
+        highscoreTable.clear();
+
+        highScores = Save.hsd.get(levelSelected).getHighScores();
+        names = Save.hsd.get(levelSelected).getNames();
+
+
+        //Add highscores to table
+        for (int i = 0; i < highScores.length; i++) {
+            Label num = new Label(String.format("%2d.", i + 1), skin, "highscore");
+            Label score = new Label(String.format("%7s", highScores[i]), skin, "highscore");
+            Label name = new Label(String.format("%s", names[i].split("-")[0]), skin, "highscore");
+
+            highscoreTable.add(num).padRight(200).right();
+            highscoreTable.add(name).padRight(90).left();
+            highscoreTable.add(score).left();
+
+            highscoreTable.row();
+        }
+
+        //highscorePane.setActor(innerHighScoreTable);
+
     }
 
     @Override
@@ -229,9 +239,9 @@ public class LevelSelectionScreen implements Screen {
 
     @Override
     public void dispose() {
-        stage.dispose();
+
     }
-
-
 }
+
+
